WIP: Allow connectors to specify types
This commit removes a need to define a global Message enum type and
allows Models to handle any event type.

It doesn't compile yet though:
- Didn't work out lifetimes fully

It also doesn't do all that I want it to:
- Models should allow generic inputs and outputs - e.g. a buffer model
  should allow for type arg T which gets stored in a queue, not sure how
  that works out with the current system.

What IS handled is type erasure of model inputs and outputs, which
prevents having to specialize a lot of structures used by the simulation
to end-user defined Message types. This makes the library defined model
components easier to use in conjunction with multiple crates defining
their own models.

Signed-off-by: Tin Svagelj <tin.svagelj@live.com>